create dataset instance from build_dataset function
